requirements.txt
click 
orjson 
attrs
pyelftools
cpp_demangle
